Переработка шапки на Nuxt UI
