+++ app/components/the-header/navigation/TheHeaderNavigation.vue
@@ -1,33 +1,12 @@
 <script setup lang="ts">
-import type { RouteName } from "~/types/Utils"
+import type { NavigationMenuItem } from "@nuxt/ui"
 
 const { navigation } = defineProps<{
   /** Навигация */
-  navigation: RouteName[]
+  navigation: NavigationMenuItem[]
 }>()
-
-const { t } = useI18n()
-const localePath = useLocalePath()
-
-const icons = {
-  about: "heroicons:user",
-  "contact-me": "heroicons:envelope",
-  experience: "heroicons:briefcase",
-  index: "heroicons:home",
-  projects: "heroicons:squares-2x2",
-} as const satisfies Record<RouteName, string>
-
-const items = computed(() => {
-  return navigation.map((routeName) => {
-    return {
-      icon: icons[routeName],
-      label: t(`pages.${routeName}`),
-      to: localePath(routeName),
-    }
-  })
-})
 </script>
 
 <template>
-  <UNavigationMenu :items />
+  <UNavigationMenu :items="navigation" />
 </template>
