+++ app/components/the-header/navigation/TheHeaderNavigationMobile.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+import type { DropdownMenuItem } from "@nuxt/ui"
+
+defineProps<{
+  /** Навигация */
+  navigation: DropdownMenuItem[]
+}>()
+</script>
+
+<template>
+  <UDropdownMenu :items="navigation" size="lg">
+    <UButton color="neutral" icon="heroicons:bars-2" size="lg" variant="soft" />
+  </UDropdownMenu>
+</template>
