+++ app/components/ui/UiContactMeButton.vue
@@ -1,16 +1,18 @@
 <script setup lang="ts">
 const { t } = useI18n({ useScope: "local" })
+
+const localePath = useLocalePath()
 </script>
 
 <template>
-  <NuxtLinkLocale class="button primary" :to="{ name: 'contact-me' }">
-    {{ t("buttonText") }}
-  </NuxtLinkLocale>
+  <UButton :label="t('buttonText')" size="lg" :to="localePath('contact-me')">
+    <template #trailing>✍🏻</template>
+  </UButton>
 </template>
 
 <i18n lang="yaml">
 en:
-  buttonText: Contact me ✍🏻
+  buttonText: Contact me
 ru:
-  buttonText: Написать мне ✍🏻
+  buttonText: Написать мне
 </i18n>
